Data layer: move verify_vcs package to "credentials"
#88
